# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Finance & Crisis Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: List all debts
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debt"
    post:
      operationId: createDebt
      tags: [debts]
      summary: Create a new debt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtInput"
      responses:
        "201":
          description: Created debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}:
    get:
      operationId: getDebt
      tags: [debts]
      summary: Get a debt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
        "404":
          description: Not found
    put:
      operationId: updateDebt
      tags: [debts]
      summary: Update a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebtInput"
      responses:
        "200":
          description: Updated debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
    delete:
      operationId: deleteDebt
      tags: [debts]
      summary: Delete a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /davlatov/allocate:
    post:
      operationId: createDavlatovAllocation
      tags: [davlatov]
      summary: Calculate and save Davlatov wealth allocation for an amount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DavlatovAllocateRequest"
      responses:
        "201":
          description: Allocation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DavlatovAllocation"
...
  /davlatov/allocations/{id}:
    delete:
      operationId: deleteDavlatovAllocation
      tags: [davlatov]
      summary: Delete a Davlatov allocation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Goals ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Debt:
      type: object
      required: [id, creditorName, totalDebt, monthlyPayment, interestRate, dueDate, createdAt]
      properties:
        id:
          type: number
        creditorName:
          type: string
        totalDebt:
          type: number
        monthlyPayment:
          type: number
        interestRate:
          type: number
          description: Annual interest rate as a percentage (e.g. 18.5 for 18.5%)
        dueDate:
          type: string
          description: ISO date string (YYYY-MM-DD)
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DebtInput:
      type: object
      required: [creditorName, totalDebt, monthlyPayment, interestRate, dueDate]
      properties:
        creditorName:
          type: string
        totalDebt:
          type: number
        monthlyPayment:
          type: number
        interestRate:
          type: number
        dueDate:
          type: string
        notes:
          type: string
...
    Expense:
      type: object
      required: [id, category, name, amount, isEssential, createdAt]
      properties:
        id:
          type: number
        category:
          type: string
          enum: [housing, food, transport, utilities, health, miscellaneous]
        name:
          type: string
        amount:
          type: number
        isEssential:
          type: boolean
        emotionalTrigger:
          type: ["string", "null"]
          enum: [routine, stress_buying, status_validation, burnout_convenience, null]
        isImpulseBuy:
          type: boolean
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, name, amount, isEssential]
      properties:
        category:
          type: string
          enum: [housing, food, transport, utilities, health, miscellaneous]
        name:
          type: string
        amount:
          type: number
        isEssential:
          type: boolean
        emotionalTrigger:
          type: string
          enum: [routine, stress_buying, status_validation, burnout_convenience]
        isImpulseBuy:
          type: boolean
...
    DavlatovAllocateRequest:
      type: object
      required: [sourceAmount, sourceType]
      properties:
        sourceAmount:
          type: number
          description: Total amount to distribute
        sourceType:
          type: string
          enum: [dividend, personal_income]
        charityPct:
          type: number
          description: Charity percentage (2.5 to 10). Defaults to 10.
        notes:
          type: string
    DavlatovAllocation:
      type: object
      required: [id, sourceType, sourceAmount, charityPct, charityAmt, parentsAmt, savingsAmt, entertainmentAmt, largeDreamAmt, smallDreamAmt, createdAt]
      properties:
        id:
          type: number
        sourceType:
          type: string
        sourceAmount:
          type: number
        charityPct:
          type: number
        charityAmt:
          type: number
        parentsAmt:
          type: number
        savingsAmt:
          type: number
        entertainmentAmt:
          type: number
        largeDreamAmt:
          type: number
        smallDreamAmt:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Goals ───────────────────────────────────────────────
...